Separate OSMP and model
