+++ src/README.md
@@ -7,5 +7,5 @@ The hello world example is based on the [OSMPDummySensor](https://github.com/Ope
 | File                      | Description                                                                                                              |
 |---------------------------|--------------------------------------------------------------------------------------------------------------------------|
 | CMakeLists.txt            | Contains the cmake instructions to build and package the model.                                                          |
-| HelloWorldSensor.cpp/.h:  | These files contain the actual model source code.                                                                        |
+| OSMP.cpp/.h:  | These files contain the actual model source code.                                                                        |
 | modelDescription.in.xml   | Contains the input for the modelDescription.xml file packaged with every FMU. The actual xml file is generated by cmake. |
